Add Reindexing Remote Tests
Adds integration tests that reindex from remote for versions 5.0, 6.2, 6
.3, 8.12, 8.16 and 9.3

Relates: elastic/elasticsearch-team#2088
